requirements.txt
mcp>=1.0.0
openai>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
kaggle>=1.5.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
asyncio>=3.9.0
aiohttp>=3.8.0